backend/requirements-test.txt
fastapi
uvicorn
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
msal
httpx
python-dotenv
pytest>=8.0.0
pytest-asyncio
openai
tiktoken
# Minimal llama-index deps for test imports (no heavy ML libs)
llama-index-core
llama-index-llms-openai-like
llama-index-embeddings-openai
llama-index-vector-stores-qdrant
qdrant-client
kfp
minio
